get volunteer mappings

+++ routes/volunteers.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const volunteersDb = require('../db/volunteers');
-const { body, param, validationResult } = require('express-validator');
+const { body, param, validationResult, query } = require('express-validator');
 
 
 // Create locations
@@ -38,7 +38,7 @@ router.put('/:volunteerId',
         }
         try {
             const volunteer = await volunteersDb.updatedEmail(req.params.volunteerId, req.body.email);
-            res.status(201).json(volunteer);
+            res.status(200).json(volunteer);
         } catch (error) {
             let errorString = `ERROR OCCURRED CREATING Volunteer! body: ${req.body} error: ${error}`;
             console.log(errorString);
@@ -47,14 +47,34 @@ router.put('/:volunteerId',
         }
     });
 
-// Create volunteer
+// Get all volunteer
 router.get('/',
     async function(req, res) {
         try {  
             const volunteers = await volunteersDb.getAllVolunteers();
-            res.status(201).json(volunteers);
+            res.status(200).json(volunteers);
         } catch (error) {
             let errorString = `ERROR OCCURRED LOOKING UP LOCATIONS! error: ${error}`;
+            console.log(errorString);
+            let errorObj = {error: errorString};
+            res.status(500).send(errorObj);
+        }
+    });
+
+
+// Get all volunteer
+router.get('/by-email',
+    query('volunteerEmail').isEmail(),
+    async function(req, res) {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
+        }
+        try {  
+            const volunteers = await volunteersDb.getVolunteerByEmail(req.query.volunteerEmail);
+            res.status(200).json(volunteers);
+        } catch (error) {
+            let errorString = `ERROR OCCURRED LOOKING UP VOLUNTEER! error: ${error}`;
             console.log(errorString);
             let errorObj = {error: errorString};
             res.status(500).send(errorObj);
@@ -64,9 +84,13 @@ router.get('/',
 // Delete volunteer
 router.delete('/:volunteerId', param('volunteerId').isNumeric(),
     async function(req, res) {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({ errors: errors.array() });
+        }
         try {  
             const volunteers = await volunteersDb.deleteVolunteer(req.params.volunteerId);
-            res.status(201).json(volunteers);
+            res.status(200).json(volunteers);
         } catch (error) {
             let errorString = `ERROR OCCURRED LOOKING UP LOCATIONS! error: ${error}`;
             console.log(errorString);
